% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_output}
\alias{bind_output}
\title{Bind outputs from a list of lists}
\usage{
bind_output(list_of_lists, type)
}
\arguments{
\item{list_of_lists}{A list of lists containing outputs}

\item{type}{The type of output to extract and bind}
}
\value{
A combined matrix of outputs
}
\description{
This function binds outputs of a specific type from a list of lists.
}
