% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_windows.R
\name{create_windows}
\alias{create_windows}
\title{Create train/test rolling windows for model evaluation}
\usage{
create_windows(
  df,
  initial_size,
  window_size,
  horizon_size,
  step_size = NULL,
  type = "expanding"
)
}
\arguments{
\item{df}{dataframe of data you want to evaluate}

\item{initial_size}{Int, How large the starting window is.}

\item{window_size}{Int, Width of sliding window.}

\item{horizon_size}{Int,  Number of values to be forecast ahead.}

\item{step_size}{Int, Number of rows added in each iteration.}

\item{type}{str, "expanding" or "sliding"}
}
\value{
df with added attributes for training and testing indexes
}
\description{
Create train/test rolling windows for model evaluation
}
