% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{eta_to_alpha}
\alias{eta_to_alpha}
\title{Convert eta parameters to alpha (weight) parameters}
\usage{
eta_to_alpha(list_of_eta)
}
\arguments{
\item{list_of_eta}{A list of eta parameters}
}
\value{
A matrix of alpha parameters (weights) that sum to 1 across rows
}
\description{
This function converts a list of eta parameters to alpha parameters using the softmax function.
The transformation ensures that the resulting weights sum to 1.
}
