% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{eta_to_beta}
\alias{eta_to_beta}
\title{Convert eta derivatives to beta derivatives}
\usage{
eta_to_beta(list_of_ll_eta, list_of_X_eta, elementwise = FALSE)
}
\arguments{
\item{list_of_ll_eta}{A list of derivatives with respect to eta}

\item{list_of_X_eta}{A list of design matrices}

\item{elementwise}{Logical; if TRUE, performs elementwise multiplication instead of matrix multiplication}
}
\value{
A list of derivatives with respect to beta
}
\description{
This function converts derivatives with respect to eta to derivatives with respect to beta
by applying the chain rule.
}
