% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{etaeta_to_betabeta}
\alias{etaeta_to_betabeta}
\title{Convert second derivatives with respect to eta to second derivatives with respect to beta}
\usage{
etaeta_to_betabeta(ll_etaeta, list_of_X_eta, k1, k2)
}
\arguments{
\item{ll_etaeta}{A matrix of second derivatives with respect to eta}

\item{list_of_X_eta}{A list of design matrices}

\item{k1}{Index of the first eta parameter}

\item{k2}{Index of the second eta parameter}
}
\value{
A matrix of second derivatives with respect to beta
}
\description{
This function converts second derivatives (Hessian) with respect to eta to second derivatives
with respect to beta by applying the chain rule.
}
