% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{eval_deriv}
\alias{eval_deriv}
\title{Evaluates a set of inner derivatives on given parameters}
\usage{
eval_deriv(inner_deriv_object, eta, theta, deriv = 0)
}
\arguments{
\item{inner_deriv_object}{Output of any inner function}

\item{eta}{Vector of linear predictor values}

\item{theta}{Vector of extra parameters}

\item{deriv}{integer dictating what level of derivatives are desired}
}
\value{
Named list of derivatives
}
\description{
Evaluates a set of inner derivatives on given parameters
}
