% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts.R
\name{evaluate_all}
\alias{evaluate_all}
\title{Fit a set of experts and stack them over an out-of-sample window.}
\usage{
evaluate_all(experts, stacker, formula, expert_windower, stack_windower, dat)
}
\arguments{
\item{experts}{description}

\item{stacker}{Stacker object created by \code{create_stacker}.}

\item{formula}{Model formula for the weights. For nested stucture this should be a list of lists.}

\item{expert_windower}{A windower object created by \code{create_windower} used for evaluating experts.}

\item{stack_windower}{A windower object created by \code{create_windower} used for evaluating stacks.}

\item{dat}{Data that should be used for fitting experts and stacking, should contain all columns specified by formula.}
}
\value{
list of predicted weights from stacking process.
}
\description{
Fit a set of experts and stack them over an out-of-sample window.
}
