% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts.R
\name{evaluate_expert}
\alias{evaluate_expert}
\title{Evaluate an expert object over a sliding or expanding window.}
\usage{
evaluate_expert(expert, windower, data, type = "predict")
}
\arguments{
\item{expert}{An expert object created by \code{create_expert}.}

\item{windower}{A window object created by \code{create_windower}.}

\item{data}{A dataframe containing response and features.}

\item{type}{List the options you want outputted from "predict", "density" and "model".}
}
\value{
A list containing the outputs determined by type.
}
\description{
Evaluate an expert object over a sliding or expanding window.
}
