% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts.R
\name{evaluate_stack}
\alias{evaluate_stack}
\title{Evaluate a set of experts with weighting functions.}
\usage{
evaluate_stack(stacker, formula, windower, stack_data, preds, RidgePen = 1e-05)
}
\arguments{
\item{stacker}{Stacker object created by \code{create_stacker}.}

\item{formula}{Model formula for the weights. For nested stucture this should be a list of lists.}

\item{windower}{A windower object created by \code{create_windower}.}

\item{stack_data}{Data that should be used for stacking, should contain all columns specified by formula.}

\item{list_of_densities}{For nested stack this is a list of lists. Densities can be retrieved from experts. Rows should match that of \code{stack_data}.}
}
\value{
list of predicted weights from stacking process.
}
\description{
Evaluate a set of experts with weighting functions.
}
