% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fr_electricity_price}
\alias{fr_electricity_price}
\title{French Electricity Spot Prices}
\format{
A data frame with 34896 rows and 113 variables:
\describe{
\item{Date}{Date of the observation}
\item{Hour}{Hour of the observation}
\item{SpotPrice}{Spot price in EUR/MWh}
}
}
\source{
\url{https://www.entsoe.eu/}

\url{https://odre.opendatasoft.com/explore/dataset/eco2mix-national-cons-def}
}
\usage{
fr_electricity_price
}
\description{
Contains historical spot prices for electricity in France as well as generation mix and consumption data.
}
\keyword{datasets}
