% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_nestedstack.R
\name{get_derivatives}
\alias{get_derivatives}
\title{get_derivatives}
\usage{
get_derivatives(
  list_of_beta,
  list_of_betaT,
  list_of_theta,
  list_of_X_eta,
  list_of_X_etaT,
  list_of_inner_functions,
  list_of_densities,
  derivs = 1
)
}
\arguments{
\item{derivs}{}
}
\description{
get_derivatives
}
