% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_diag}
\alias{get_diag}
\title{Extract diagonal elements from a list of lists}
\usage{
get_diag(list_of_lists)
}
\arguments{
\item{list_of_lists}{A list of lists}
}
\value{
A list containing the diagonal elements
}
\description{
This function extracts the diagonal elements from a list of lists.
}
