% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_eval}
\alias{get_eval}
\title{Retrieve evaluated derivatives from storage object}
\usage{
get_eval(eval_string, eval_store)
}
\arguments{
\item{eval_string}{String for name of derivative you want}

\item{eval_store}{A storage object created by eval_deriv}
}
\value{
A list of evaluated derivatives
}
\description{
Retrieve evaluated derivatives from storage object
}
