% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_list_of_eta}
\alias{get_list_of_eta}
\title{Calculates eta for given design matrices and coefficients}
\usage{
get_list_of_eta(list_of_X_eta, list_of_beta)
}
\arguments{
\item{list_of_X_eta}{List of design matrices for eta}

\item{list_of_beta}{List of coefficients for eta}
}
\value{
A list of eta vectors
}
\description{
Calculates eta for given design matrices and coefficients
}
