% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_DensStack.R
\name{get_ll_dens_derivs}
\alias{get_ll_dens_derivs}
\title{Calculate log-likelihood derivatives for density stacking}
\usage{
get_ll_dens_derivs(
  list_of_beta,
  list_of_X,
  theta,
  weight,
  log_dens,
  exp_log_dens_maxs,
  exp_log_dens_diff,
  deriv = 0
)
}
\arguments{
\item{list_of_beta}{List of beta parameters}

\item{list_of_X}{List of design matrices}

\item{theta}{Theta parameters}

\item{weight}{Weight function}

\item{log_dens}{Matrix of log densities}

\item{exp_log_dens_maxs}{Maximum exponentiated log densities}

\item{exp_log_dens_diff}{Exponentiated differences from maximum log densities}

\item{deriv}{Integer indicating derivative order (0=value, 1=gradient, 2=Hessian)}
}
\value{
List containing log-likelihood (l), gradient (lb), and Hessian (lbb)
}
\description{
This function calculates the log-likelihood and its derivatives for density stacking models.
It computes the gradient and Hessian matrix needed for model fitting.
}
