% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_lossstack.R
\name{get_loss_derivs}
\alias{get_loss_derivs}
\title{Calculate loss function derivatives for stacking}
\usage{
get_loss_derivs(
  list_of_beta,
  list_of_X,
  theta,
  logv,
  loss,
  weight,
  preds,
  y,
  deriv
)
}
\arguments{
\item{list_of_beta}{List of beta parameters}

\item{list_of_X}{List of design matrices}

\item{theta}{Theta parameters}

\item{logv}{Log of the variance parameter}

\item{loss}{Loss function}

\item{weight}{Weight function}

\item{preds}{Matrix of predictions from different models}

\item{y}{Response variable}

\item{deriv}{Integer indicating derivative order (0=value, 1=gradient, 2=Hessian)}
}
\value{
List containing log-likelihood (l), gradient (lb), and Hessian (lbb)
}
\description{
This function calculates the derivatives of the loss function for stacking models.
It computes the gradient and Hessian matrix needed for model fitting.
}
