% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{id}
\alias{id}
\title{Produces equal weights without any dependence on parameters}
\usage{
id()
}
\value{
Object of derivatives for equally weighted models
}
\description{
Produces equal weights without any dependence on parameters
}
\examples{
id()
}
