% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{Latent}
\alias{Latent}
\title{Latent weights aim to recreate multinomial with less linear predictors by
using 'latent' linear predictors}
\usage{
Latent(K, q)
}
\arguments{
\item{K}{integer - Number of experts to weight}

\item{q}{integer - Number of latent factors to use}
}
\description{
Latent weights aim to recreate multinomial with less linear predictors by
using 'latent' linear predictors
}
