% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_functions.R
\name{link_theta}
\alias{link_theta}
\title{Apply link function to transform parameters}
\usage{
link_theta(f, link_func = "log", eta, tau)
}
\arguments{
\item{f}{Function to apply}

\item{link_func}{String specifying the link function to use (default: "log")}

\item{eta}{Linear predictor values}

\item{tau}{Parameter values on the link scale}
}
\value{
Transformed parameter values
}
\description{
This function applies a link function to transform parameters.
}
