% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_by_list}
\alias{list_by_list}
\title{Multiply elements of two lists}
\usage{
list_by_list(list1, list2, elementwise = FALSE)
}
\arguments{
\item{list1}{First list of elements}

\item{list2}{Second list of elements}

\item{elementwise}{Logical; if TRUE, performs elementwise multiplication instead of matrix multiplication}
}
\value{
A list containing the products
}
\description{
This function multiplies corresponding elements of two lists, either elementwise
or using matrix multiplication.
}
