% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_by_vector}
\alias{list_by_vector}
\title{Multiply each element of a list by corresponding element of a vector}
\usage{
list_by_vector(l, v)
}
\arguments{
\item{l}{A list of elements}

\item{v}{A vector of multipliers}
}
\value{
A list containing the products
}
\description{
This function multiplies each element of list l by the corresponding element of vector v.
}
