% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_take_list}
\alias{list_take_list}
\title{Subtract elements of one list from another}
\usage{
list_take_list(l1, l2)
}
\arguments{
\item{l1}{First list}

\item{l2}{Second list (to be subtracted from l1)}
}
\value{
A list containing the differences between corresponding elements
}
\description{
This function subtracts corresponding elements of list l2 from list l1.
}
