% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_times_list}
\alias{list_times_list}
\title{Elementwise multiplication of two lists}
\usage{
list_times_list(list1, list2)
}
\arguments{
\item{list1}{First list of elements}

\item{list2}{Second list of elements}
}
\value{
A list containing the elementwise products
}
\description{
This function performs elementwise multiplication of corresponding elements in two lists.
}
