% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gaussian.R
\name{llk_gaussian2}
\alias{llk_gaussian2}
\title{Calculate Gaussian log-likelihood and its derivatives (matrix version)}
\usage{
llk_gaussian2(y, param, deriv = 0, ...)
}
\arguments{
\item{y}{Matrix of observations}

\item{param}{List or matrix of parameters: first column is mean, second column is precision (1/sigma)}

\item{deriv}{Integer indicating derivative order (0=value, 1=gradient, 2=Hessian)}

\item{...}{Additional arguments (not used)}
}
\value{
A list containing the log-likelihood and its derivatives
}
\description{
This function calculates the log-likelihood of a Gaussian distribution and its derivatives
with respect to the mean and precision parameters. It's optimized for matrix inputs.
}
