% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{log1px}
\alias{log1px}
\title{Numerically stable computation of log(1+exp(x))}
\usage{
log1px(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector containing log(1+exp(x))
}
\description{
This function computes log(1+exp(x)) in a numerically stable way by using different
approximations depending on the range of x.
}
