% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_cumsum}
\alias{mat_cumsum}
\title{Cumulative sum of a list of matrices}
\usage{
mat_cumsum(list_of_mats)
}
\arguments{
\item{list_of_mats}{A list of matrices}
}
\value{
A list of cumulative sums of matrices
}
\description{
This function calculates the cumulative sum of a list of matrices.
}
