% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat_split}
\alias{mat_split}
\title{Split a matrix into sub-matrices}
\usage{
mat_split(M, r, c)
}
\arguments{
\item{M}{Input matrix}

\item{r}{Number of rows in each sub-matrix}

\item{c}{Number of columns in each sub-matrix}
}
\value{
An array of sub-matrices
}
\description{
This function splits a matrix into sub-matrices of specified dimensions.
}
