% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_to_lov}
\alias{matrix_to_lov}
\title{Convert a matrix to a list of vectors}
\usage{
matrix_to_lov(X)
}
\arguments{
\item{X}{Input matrix}
}
\value{
A list of column vectors
}
\description{
This function converts a matrix to a list of column vectors.
}
