% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{max_index_mat}
\alias{max_index_mat}
\title{Matrix of max(i,j) where i,j are matrix coords}
\usage{
max_index_mat(n)
}
\arguments{
\item{n}{Integer: Size of n x n matrix}
}
\value{
Matrix of size n x n
}
\description{
Matrix of max(i,j) where i,j are matrix coords
}
