% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{multinomial}
\alias{multinomial}
\title{Multinomial weights}
\usage{
multinomial(K, normalised = TRUE)
}
\arguments{
\item{K}{\itemize{
\item Integer specifying how many weights are used
}}

\item{normalised}{\itemize{
\item Boolean, should weights be normalised
}}
}
\value{
\itemize{
\item Weight object
}
}
\description{
Multinomial weights
}
