% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{MVN_weights}
\alias{MVN_weights}
\title{Produces an object containing derivatives for multivariate normal stacking}
\usage{
MVN_weights(x, dim_num)
}
\arguments{
\item{x}{Array of (dim_num, n_k) specifying locations for each of the models}

\item{dim_num}{Integer specifying number of dimensions required}
}
\value{
Object containing derivatives for multivariate normal stacking
}
\description{
Produces an object containing derivatives for multivariate normal stacking
}
\examples{
x <- matrix(c(1:6), nrow = 3)
MVN_weights(x, 3)
}
