% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{MVN_weights_cpp}
\alias{MVN_weights_cpp}
\title{Produces an object containing derivatives for multivariate normal stacking
Uses cpp for faster computation}
\usage{
MVN_weights_cpp(ex_coords)
}
\arguments{
\item{ex_coords}{Array of (dim_num, K) specifying locations for each of the experts, dim_num is dimension of expert space}
}
\value{
Object containing derivatives for multivariate normal stacking
}
\description{
Produces an object containing derivatives for multivariate normal stacking
Uses cpp for faster computation
}
\examples{
x <- matrix(c(1:6), nrow = 3)
MVN_weights(x, 3)
}
