% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{nested}
\alias{nested}
\title{Nested weights, weights a set of inner weighting functions by function given
in outer weight}
\usage{
nested(outer_weight, inner_weight)
}
\arguments{
\item{outer_weight}{\itemize{
\item Weighting function to multiply inner weights
}}

\item{inner_weight}{\itemize{
\item List of inner weighting functions
}}
}
\value{
A weighting object that multiplies set of outer weights by inner weights
}
\description{
Nested weights, weights a set of inner weighting functions by function given
in outer weight
}
