% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_nestedstack.R
\name{NestedStack}
\alias{NestedStack}
\title{mgcv family for nested stacking}
\usage{
NestedStack(P, inner_funcs, RidgePen = 1e-05)
}
\arguments{
\item{P}{List of Matrices of model densities}

\item{inner_funcs}{List of inner function objects}

\item{RidgePen}{Positive numeric the scalar applied to the ridge penalty in regression}
}
\value{
An initialised mgcv family object
}
\description{
mgcv family for nested stacking
}
\examples{
inners = list(ordinal(3), id(1, 100))
P = list(matrix(rnorm(300), nrow = 100), matrix(rnorm(100), nrow = 100))
NestedStack(P, inners)
}
