% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{ordinal}
\alias{ordinal}
\title{Produces an object containing derivatives for ordered stacking}
\usage{
ordinal(num_weights)
}
\arguments{
\item{num_weights}{Integer: How many weights in the ordering}
}
\value{
List of functions for finding ordinal weights
}
\description{
Produces an object containing derivatives for ordered stacking
}
\examples{
ordinal(3)
}
