% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Create a pinball loss function for quantile regression}
\usage{
pinball_loss(tau)
}
\arguments{
\item{tau}{Quantile level (between 0 and 1)}
}
\value{
A function that calculates the pinball loss and its derivatives
}
\description{
This function creates a pinball loss function for a specified quantile level.
}
