% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_windows.R
\name{plot_windows}
\alias{plot_windows}
\title{Plot windows for easy visualisation.}
\usage{
plot_windows(windowed_dat)
}
\arguments{
\item{windowed_dat}{Output of create_windows.}
}
\value{
Displays plot showing windowed data
}
\description{
Plot windows for easy visualisation.
}
