% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repmat}
\alias{repmat}
\title{Replicate a matrix}
\usage{
repmat(mat, v = 1, h = 1)
}
\arguments{
\item{mat}{Input matrix}

\item{v}{Number of vertical replications}

\item{h}{Number of horizontal replications}
}
\value{
A replicated matrix
}
\description{
This function replicates a matrix vertically and horizontally.
}
