% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_windows.R
\name{split_by}
\alias{split_by}
\title{Split a dataframe by a factor variable}
\usage{
split_by(df, fact_split)
}
\arguments{
\item{df}{Dataframe to split}

\item{fact_split}{Column name of the factor variable to split by}
}
\value{
A list of dataframes, one for each unique value of the factor variable
}
\description{
This function splits a dataframe into multiple dataframes based on unique values of a factor variable.
}
