% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts.R
\name{create_expert}
\alias{create_expert}
\title{Create an expert that can be fitted, retrieve predictions and simulate from fitted distribution.}
\usage{
create_expert(fit_func, pred_func = NULL, sim_func = NULL, dens_func = NULL)
}
\arguments{
\item{fit_func}{A function that returned a fitted model given a dataframe.}

\item{pred_func}{A function that returns predictions given fitted model and data.}

\item{sim_func}{A function that simulated from the estimated density given model and data.}

\item{dens_func}{A function returns density under fitted model at given data points.}
}
\value{
An expert object that can be used within evaluate_expert and evaluate_stack.
}
\description{
Create an expert that can be fitted, retrieve predictions and simulate from fitted distribution.
}
