% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experts.R
\name{create_stacker}
\alias{create_stacker}
\title{Create a stacking object from experts and inner functions.}
\usage{
create_stacker(experts, inners)
}
\arguments{
\item{experts}{List of expert objects created by \code{create_expert}.}

\item{inners}{Function that determines how weights}
}
\value{
Stacking object that can take data and return predicted weights for experts.
}
\description{
Create a stacking object from experts and inner functions.
}
