% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_functions.R
\name{id}
\alias{id}
\title{Produces equal weights without any dependence on parameters}
\usage{
id()
}
\arguments{
\item{num_weights}{Integer: How many equally weighted parts}

\item{N}{Integer: Number of data points}
}
\value{
Object of derivatives for equally weighted models
}
\description{
Produces equal weights without any dependence on parameters
}
\examples{
id(3, 1000)
}
